Wait for all commands to start before proxying signals
There's a race condition in child process setup and signal handling -
if we get a signal, ProxySignals iterates over the children and calls
t.Process.Signal(sig).  If a signal comes in right after ProxySignals
is called but before cmd.Run() returns in any of the goroutines in
Codep, t.Process will be nil and a segfault will occur.

This change uses cmd.Start() in the main goroutine to make sure all
children are setup (or have otherwise errored out) before setting up
signal handling.